% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.r
\name{AICc}
\alias{AICc}
\title{AICc from gam or lm object}
\usage{
AICc(object)
}
\arguments{
\item{object}{A model returned by a gam or lm fit}
}
\value{
an AICc value
}
\description{
Helper function compute AICc from a gam or lm object
}
\examples{
# some simulated data with one cov
dat <- mgcv::gamSim(6,n=20,scale=.5)[,1:2]
m <- mgcv::gam(y~s(x0), data=dat)
AICc(m)
}
