% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc_gam.r
\name{AICcgam}
\alias{AICcgam}
\title{AICc from gam object}
\usage{
AICcgam(object)
}
\arguments{
\item{object}{A model returned by a gam fit}
}
\value{
an AICc value
}
\description{
Helper function compute AICc from a gam object
}
\examples{
# some simulated data with one cov
dat <- mgcv::gamSim(6,n=20,scale=.5)[,1:2]
m <- mgcv::gam(y~s(x0), data=dat)
AICcgam(m)
}
