% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-helper.R
\name{as_datatable}
\alias{as_datatable}
\title{Display an object as a DT datatable}
\usage{
as_datatable(x, ...)
}
\arguments{
\item{x}{Object to display (data frame, table_one, report, etc.)}

\item{...}{Additional arguments passed to DT::datatable}
}
\value{
DT datatable (in Viewer)
}
\description{
Generic helper to show any supported sardine object as an interactive DT table.
Works for data frames, table_one, reports, metadata, etc.
}
