% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chl}
\alias{chl}
\title{Chlorophyll-a data from remote-sensing products}
\format{
A data frame with:
\describe{
  \item{Year}{The year}
  \item{Month}{The month}
  \item{CHL.1 to CHL.13}{Average monthly CHL averaged over boxes 1 to 13. Units are mg/m$^3$.}
}
}
\usage{
chl
}
\description{
The CHL satellite data (mg/m$^3$) were downloaded from the NOAA ERDDAP server using 
Roy Mendelssohn's
 \strong{rerddapXtracto} R package which uses the ropensci\strong{rerddap} R packages available on CRAN.  The R code used
to download the data is in the \code{extdata/get_satelite_data} folder.  See examples for how to find
the file.
}
\details{
The Chlorophyll-a products are developed by the
Ocean Biology Processing Group in the Ocean Ecology Laboratory
at the NASA Goddard Space Flight Center.

For September 1997 to 2002, we used the Chlorophyll-a 2018.0 Reprocessing (R2018.0) product 
from the Sea-viewing Wide Field-of-view 
Sensor (SeaWiFS) on the Orbview-2 satellite. These data are on a 0.083 degree grid.  See reference below.
   
For 2003 to 2017, we used the MODIS-Aqua R2018.1 product on a 0.0416 degree. These CHL data are taken from measurements 
gathered by the Moderate 
Resolution Imaging Spectroradiometer (MODIS) 
on NASA's Aqua Spacecraft. See reference below.
   
Both CHL data sets were downloaded from the NOAA ERDDAP server. See
    \url{https://coastwatch.pfeg.noaa.gov/erddap/info/erdSW2018chlamday/index.html} and 
    \url{https://coastwatch.pfeg.noaa.gov/erddap/info/erdMH1chlamday/index.html}.
   
The R code used
to download the data is in the \code{extdata/get_satelite_data} folder.  See examples
for how to find and view the files.  The CHL values were averaged across thirteen
1 degree by 1 degree boxes 
which roughly parallel the bathymetry.  See the figure in the examples below which shows
the boxes.
}
\examples{
\dontrun{
# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_sat_data.R", package="SardineForecast"))

# Show the background files on each data set at the time the data were downloaded
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdSW1chlamday", package="SardineForecast"))
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdMH1chlamday", package="SardineForecast"))

# Show the boxes
browseURL(system.file("docs", "kerala_study_area_with_inset.jpg", package="SardineForecast"))
}
}
\references{
NASA Goddard Space Flight Center, Ocean Ecology Laboratory, 
Ocean Biology Processing Group; (2018): SeaWiFS Ocean Color Data; 
NASA Goddard Space Flight Center, Ocean Ecology Laboratory, 
Ocean Biology Processing Group. \url{https://dx.doi.org/10.5067/ORBVIEW-2/SEAWIFS/L3B/CHL/2018}

NASA Goddard Space Flight Center, Ocean Ecology Laboratory, Ocean 
Biology Processing Group. Moderate-resolution Imaging 
Spectroradiometer (MODIS) Aqua Chlorophyll Data; 2014 Reprocessing. 
NASA OB.DAAC, Greenbelt, MD, USA. \url{https://dx.doi.org/10.5067/AQUA/MODIS/L3M/CHL/2018}


\insertRef{Huetal2012}{SardineForecast}

\insertRef{rerddap}{SardineForecast}

\insertRef{rerddapXtracto}{SardineForecast}
}
\keyword{datasets}
