% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{.build_overall_table}
\alias{.build_overall_table}
\title{Build Overall Summary Table}
\usage{
.build_overall_table(
  data,
  vars,
  var_types,
  non_normal,
  include_missing,
  digits
)
}
\description{
Internal function to build unstratified table
}
\keyword{internal}
