% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{.detect_variable_types}
\alias{.detect_variable_types}
\title{Detect Variable Types}
\usage{
.detect_variable_types(data, metadata, vars, categorical, continuous)
}
\description{
Internal function to detect whether variables are categorical or continuous
}
\keyword{internal}
