% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-summary-stats.R
\name{.duration_efficiency_regularity}
\alias{.duration_efficiency_regularity}
\title{Duration, efficiency, regularity summary}
\usage{
.duration_efficiency_regularity(summary)
}
\arguments{
\item{summary}{processed summary tibble}
}
\value{
tibble with aggregated metrics
}
\description{
Duration, efficiency, regularity summary
}
\keyword{internal}
