% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-epoch-plots.R
\name{.generate_nightly_epoch_plots}
\alias{.generate_nightly_epoch_plots}
\title{Generate nightly epoch plots for a single sleep session}
\usage{
.generate_nightly_epoch_plots(sleep_id, epoch_data, summary_data)
}
\arguments{
\item{sleep_id}{The sleep ID to generate plots for}

\item{epoch_data}{The epoch data frame}

\item{summary_data}{The summary data frame}
}
\value{
A list of plotly plot objects
}
\description{
Generate nightly epoch plots for a single sleep session
}
\keyword{internal}
