% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-epoch-plots.R
\name{.plot_epoch_line}
\alias{.plot_epoch_line}
\title{Plot epoch line chart (HR, RR, etc.)}
\usage{
.plot_epoch_line(data, column, label, timezone)
}
\description{
Plot epoch line chart (HR, RR, etc.)
}
\keyword{internal}
