% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-epoch-plots.R
\name{.plot_hypnogram}
\alias{.plot_hypnogram}
\title{Plot hypnogram (sleep stages over time)}
\usage{
.plot_hypnogram(data, timezone)
}
\description{
Plot hypnogram (sleep stages over time)
}
\keyword{internal}
