% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-utils.R
\name{.redcap_request}
\alias{.redcap_request}
\title{Internal function to make REDCap API requests}
\usage{
.redcap_request(connection, content, format = "json", ...)
}
\arguments{
\item{connection}{A redcap_connection object}

\item{content}{Character string specifying the content type}

\item{format}{Character string specifying the response format}

\item{...}{Additional parameters to pass to the API}
}
\value{
httr2 response object
}
\description{
Internal function to make REDCap API requests
}
\keyword{internal}
