% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-summary-stats.R
\name{.sleep_ritual}
\alias{.sleep_ritual}
\title{Sleep ritual summary}
\usage{
.sleep_ritual(summary)
}
\arguments{
\item{summary}{processed summary tibble}
}
\value{
tibble with ritual metrics (bedtime, wake-up, etc.)
}
\description{
Sleep ritual summary
}
\keyword{internal}
