% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-summary-stats.R
\name{.sleep_vitals}
\alias{.sleep_vitals}
\title{Sleep vitals summary}
\usage{
.sleep_vitals(summary)
}
\arguments{
\item{summary}{processed summary tibble}
}
\value{
tibble with AHI, snoring, HR stats
}
\description{
Sleep vitals summary
}
\keyword{internal}
