% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{.summarize_categorical}
\alias{.summarize_categorical}
\title{Summarize Categorical Variable}
\usage{
.summarize_categorical(x, var_name, include_missing)
}
\description{
Internal function to compute n (\%) for categorical variables
}
\keyword{internal}
