% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{.test_categorical}
\alias{.test_categorical}
\title{Test Categorical Variable Across Strata}
\usage{
.test_categorical(x, strata)
}
\description{
Internal function to perform chi-square or Fisher's exact test
}
\keyword{internal}
