% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{.test_continuous}
\alias{.test_continuous}
\title{Test Continuous Variable Across Strata}
\usage{
.test_continuous(x, strata, test_type, non_normal)
}
\description{
Internal function to perform t-test, ANOVA, or Kruskal-Wallis test
}
\keyword{internal}
