% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enso}
\alias{enso}
\title{ENSO Indices}
\format{A data frame with variables:
\describe{
  \item{Year}{}
  \item{Month}{This is a 3-month average.  Jan value is average of Dec-Jan-Feb.  Qtr would be denoted JFM.}
  \item{ONI}{The ONI index value}
  \item{SOI}{The SOI index value}
}}
\usage{
enso
}
\description{
Oceanic Nino Index and Equatorial Southern Oscillation Index
}
\details{
The ONI index is 3 month running mean of ERSST.v5 SST anomalies in the Niño 3.4 region (5oN-5oS, 120o-170oW)], based on centered 30-year base periods updated every 5 years.  The ONI was downloaded as follows:
\preformatted{
require(XML)
require(RCurl)
require(httr)
oni=read.table("http://www.cpc.ncep.noaa.gov/data/indices/oni.ascii.txt", stringsAsFactors=FALSE)
}
The SOI used is the Equatorial SOI monthly.  The data were downloaded with the following code.
\preformatted{
soi=read.table("http://www.cpc.ncep.noaa.gov/data/indices/reqsoi.for", stringsAsFactors=FALSE)
}
}
\keyword{datasets}
