% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-shiny.R
\name{explore_project_shiny}
\alias{explore_project_shiny}
\title{Explore REDCap Project in a Shiny App}
\usage{
explore_project_shiny(project)
}
\arguments{
\item{project}{A redcap_project object}
}
\description{
Launches a Shiny app to interactively explore project structure, metadata, and data.
}
