% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{generate_quality_report_markdown}
\alias{generate_quality_report_markdown}
\title{Generate Quality Report Markdown}
\usage{
generate_quality_report_markdown(report)
}
\arguments{
\item{report}{A data_quality_report object}
}
\value{
Character vector of markdown lines
}
\description{
Internal function to generate markdown version of quality report
}
\keyword{internal}
