% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged.r
\name{lagged}
\alias{lagged}
\title{Create a k-lagged time series}
\usage{
lagged(x, k)
}
\arguments{
\item{x}{a vector}

\item{k}{the lag to use}
}
\value{
A lagged version of the vector
}
\description{
Helper function to create a lagged version of a vector.
}
\examples{
lagged(1:10, 1)
lagged(1:10, -1)
lagged(1:10, 0)
}
