% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker2qs.r
\name{maker2qs}
\alias{maker2qs}
\title{Get R2 values for linear regression of log catch anomalies vs covariates}
\usage{
maker2qs(region = "Kerala", covname = "Precip", boxes = 1:14,
  rm.trend = TRUE, rm.season = TRUE)
}
\arguments{
\item{region}{region ("Kerala", "Karnataka", "Goa") in quotes.}

\item{covname}{name of the covariate.  Must be a colname in seio_covariates_qtr.  If one of the covariates
is from the boxes, then specify the part of the name in front of the number.  For example "SST.".}

\item{boxes}{for covariate from a box, the box number or numbers to use}

\item{rm.trend}{remove year effect from catch and covariates}

\item{rm.season}{remove season effect from catch and covariates}
}
\value{
A data.frame of R2's for each lag 0 to 10.
}
\description{
This function gets the R2 values for a linear regression of 
log catch anomaly ~ covariate anomalies at lags of 0 to 10. Covariates are turned into 
anomalies by removing year and qtr effect and then lagged and z-scored.
Catch anomalies are logged and z-scored.  Note: this is an old function and not used after I
switched to using GAMs.
}
\examples{
maker2qs(region="Kerala", covname="SST.", boxes=5)

}
\keyword{internal}
