% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oilsardine_yr}
\alias{oilsardine_yr}
\title{Oil sardine landings in metric tons by quarter}
\format{A data frame 'oilsardine_qtr' with 6 variables:
\describe{
  \item{Year}{Year}
  \item{Goa}{Goa landings}
  \item{Karnataka}{Karnataka landings}
  \item{Kerala}{Kerala landings}
}}
\usage{
oilsardine_yr
}
\description{
A dataset containing the calendar year (Jan-Dec) landings (in metric tons) of oil sardines in Kerala, Karnataka and Goa states 1956-2016.  The data are collected and processed into a total landings estimates based on a stratified sampling of landing sites along the SW coast of India throughout the year.  The program is run by the Central Marine Fisheries Research Institute (CMFRI) in Cochin, India.  The data were obtained from reports published by CMFRI; see references.
}
\references{
See oilsardine_qtr for the 1956-2015 sources. 2016-present were taken from
the Fish Catch Estimates provided on the CMFRI website
}
\keyword{datasets}
