% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion-reports.R
\name{print.attrition_over_time}
\alias{print.attrition_over_time}
\title{Print Attrition Over Time}
\usage{
\method{print}{attrition_over_time}(x, ...)
}
\arguments{
\item{x}{An attrition_over_time object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Attrition Over Time
}
