% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion-reports.R
\name{print.event_completion_summary}
\alias{print.event_completion_summary}
\title{Print Event Completion Summary}
\usage{
\method{print}{event_completion_summary}(x, ...)
}
\arguments{
\item{x}{An event_completion_summary object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Event Completion Summary
}
