% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-internal.R
\name{print.redcap_project}
\alias{print.redcap_project}
\title{Print Method for REDCap Projects}
\usage{
\method{print}{redcap_project}(x, ...)
}
\arguments{
\item{x}{A redcap_project object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for REDCap Projects
}
