% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion-reports.R
\name{print.retention_summary}
\alias{print.retention_summary}
\title{Print Retention Summary}
\usage{
\method{print}{retention_summary}(x, ...)
}
\arguments{
\item{x}{A retention_summary object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Retention Summary
}
