% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-base.R
\name{print.sardine_connection}
\alias{print.sardine_connection}
\title{Print method for sardine_connection objects}
\usage{
\method{print}{sardine_connection}(x, ...)
}
\arguments{
\item{x}{A sardine_connection object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for sardine_connection objects
}
