% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{print.table_one}
\alias{print.table_one}
\title{Print Method for Table One}
\usage{
\method{print}{table_one}(x, ...)
}
\arguments{
\item{x}{A table_one object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for table_one objects
}
