% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-exports.R
\name{redcap_export_arms}
\alias{redcap_export_arms}
\title{Export REDCap Arms}
\usage{
redcap_export_arms(connection, arms = NULL, format = "json")
}
\arguments{
\item{connection}{A redcap_connection object}

\item{arms}{Character or numeric vector of specific arm numbers to export (optional)}

\item{format}{Character string. Format for returned data: "json" (default), "csv", "xml"}
}
\value{
A tibble containing arm information (when format = "json"); otherwise a character string
}
\description{
Exports the list of arms in a REDCap project.
}
