% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-exports.R
\name{redcap_export_events}
\alias{redcap_export_events}
\title{Export REDCap Events}
\usage{
redcap_export_events(connection, arms = NULL, format = "json")
}
\arguments{
\item{connection}{A redcap_connection object}

\item{arms}{Character or numeric vector of arm numbers to export events for (optional)}

\item{format}{Character string. Format for returned data: "json" (default), "csv", "xml"}
}
\value{
A tibble containing event information (when format = "json"); otherwise a character string
}
\description{
Exports the list of events in a REDCap project.
}
