% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap-exports.R
\name{redcap_export_instruments}
\alias{redcap_export_instruments}
\title{Export REDCap Instruments (Forms)}
\usage{
redcap_export_instruments(connection, format = "json")
}
\arguments{
\item{connection}{A redcap_connection object}

\item{format}{Character string. Format for returned data: "json" (default), "csv", "xml"}
}
\value{
A tibble containing instrument information (when format = "json"); otherwise a character string
}
\description{
Exports the list of instruments (forms) in a REDCap project.
}
