% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{redcap_field_names}
\alias{redcap_field_names}
\title{Get REDCap Field Names}
\usage{
redcap_field_names(connection)
}
\arguments{
\item{connection}{A redcap_connection object created by \code{\link{redcap_connection}}}
}
\value{
A character vector of field names
}
\description{
Retrieves a list of all field names in a REDCap project.
}
\examples{
\dontrun{
conn <- redcap_connection(
  url = "https://redcap.example.edu/api/",
  token = "YOUR_API_TOKEN"
)

# Get all field names
field_names <- redcap_field_names(conn)
}

}
