% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retfit.r
\name{retfit}
\alias{retfit}
\title{Fit linear regression with covariate value lagged}
\usage{
retfit(covname, catch.qtr = 1:4, lag = 0, boxes = 1:14,
  qtr.effect = FALSE, startyear = 1956, endyear = 2016)
}
\arguments{
\item{covname}{name of the covariate.  Must be a colname in seio_covariates_qtr.  If one of the covariates is from the boxes, then specify the part of the name in front of the number.}

\item{catch.qtr}{what quarter(s) of catch to use as the response}

\item{lag}{the lag specified in number of quarters. 0 means no lag.}

\item{boxes}{for covariate from a box, the box number or numbers to use}

\item{qtr.effect}{whether to make the covariate effect on catch depend on the quarter in which the catch is from}

\item{startyear}{first year}

\item{endyear}{last year}
}
\value{
An lm object of the fit
}
\description{
Fit a linear regression model where the covariate value used as explanatory variable is lagged by a 
user specified number of quarters.  The effect of a covariate can be specified to be
quarter specific, that is catch in Q1 may respond differently to the covariate than
catch in Q4. Note: this is an outdated function not used since I switched to GAMs.
}
\examples{
#Effect of precip in box 5 in qtr 1 in the previous year on catch
retfit("Precip", lag=1, boxes=5)

}
\keyword{internal}
