% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{SardineForecast-package}
\alias{SardineForecast}
\alias{SardineForecast-package}
\title{Documentation for Indian oil sardine forecasting papers}
\description{
Indian Oil Sardine Data, Code and References for Papers on Forecasting
}
\details{
To get an overview of the data, see \code{\link{precip}}, \code{\link{oilsardine_qtr}},
\code{\link{enso}}, \code{\link{chl}}, \code{\link{sst}}, \code{\link{onset}}, and
\code{\link{tidelevel}}.
To see the code for the data and raw data files, look in the \code{extdata} folder.  The full reference bib file is in \code{Sardine.bib} in the \code{docs}
along with some figures used in the papers.  \code{REFERENCES.bib} is only references for the help files.
Most of the data in the package are documented in the help files.
To see all the help files type \code{help(package="SardineForecast")}.
}
