% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sst}
\alias{sst}
\title{Sea surface temperature data from remote-sensing products}
\format{
A data frame with:
\describe{
  \item{Year}{The year}
  \item{Month}{The month}
  \item{SST.1 to SST.13}{Average monthly SST averaged over boxes 1 to 13. Units are degree Celcius.}
  \item{SSTICOAD.1 to SSTICOAD.13}{Average monthly SST averaged over boxes 1 to 13. Units are degree Celcius.}
}
}
\usage{
sst
}
\description{
The SST satellite data were downloaded from the NOAA ERDDAP server using Roy Mendelssohn's
\strong{rerddapXtracto} R package which uses the ropensci\strong{rerddap} R package available on CRAN.  The R code used
to download the data is in the \code{extdata/get_satelite_data} folder.  See examples for how to find
the file. Units are degree Celsius.
}
\details{
There are two SST data sets: AVHRR and ICOADS.

\strong{AVHRR}
The SST differential upwelling indices were computed from the SST Daily Optimum
Interpolation (OI), AVHRR Only, Version 2.1, Final 0.25 degree grid, Global data
downloaded from the NOAA ERDDAP server. These are daily data. They daily data were
averaged over all days in each month to produce monthly averages.
The SST estimates use the
Advanced Very-High Resolution Radiometer (AVHRR) instruments. AVHRR is 
accurate for close to the coast and thus could be used for our SST differential upwelling
index.
   
The SST data were downloaded from the NOAA ERDDAP server. See
    \url{https://coastwatch.pfeg.noaa.gov/erddap/info/ncdcOisst21Agg/index.html}.

\strong{ICOADS}
 The SST was extracted from the International Comprehensive Ocean-Atmosphere Data Set (ICOADS) 
 collection of surface marine data
 Data from 1960 onward were used, which are on a 1°x1° grid. 
 
The ICOADS SST data were downloaded from the NOAA ERDDAP server.
\url{https://coastwatch.pfeg.noaa.gov/erddap/info/esrlIcoads1ge/index.html}
More information on the ICOADS data are available at
\url{https://coastwatch.pfeg.noaa.gov/erddap/info/esrlIcoads1ge/index.html}
The R code used
to download the data is in the \code{extdata/get_satelite_data} folder.  See examples
for how to find and view the files.  The SST values were averaged across thirteen
1 degree by 1 degree boxes which roughly parallel the bathymetry.
}
\examples{
\dontrun{
# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_sat_data.R", package="SardineForecast"))

# Show the background files on each data set at the time the data were downloaded
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdAGsstamday.html", package="SardineForecast"))
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdPH2sstamday.html", package="SardineForecast"))

# Show the boxes
browseURL(system.file("docs", "kerala_study_area_with_inset.jpg", package="SardineForecast"))
}
}
\references{
The AVHRR data were provided by GHRSST and the US National Oceanographic Data Center. 
This project was supported in part by a grant from the NOAA Climate Data Record 
(CDR) Program for satellites. The data were downloaded from NOAA CoastWatch-West 
Coast Regional Node and Southwest Fisheries Science Center's Environmental 
Research Division. To cite these data in a paper, please follow the instructions 
in the license and at this link: 
\url{https://coastwatch.pfeg.noaa.gov/erddap/information.html#citeDataset}

The ICOADS data were provided by the NOAA/OAR/ESRL PSD, Boulder, Colorado, USA,
from their Web site at http://www.esrl.noaa.gov/psd/ (and downloaded via NOAA's
CoastWatch data server.)

\insertRef{Caseyetal2010}{SardineForecast}

\insertRef{Waltonetal1998}{SardineForecast}

\insertRef{rerddap}{SardineForecast}

\insertRef{rerddapXtracto}{SardineForecast}
}
\keyword{datasets}
