% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-base.R
\name{test_connection.default}
\alias{test_connection.default}
\title{Default test connection method}
\usage{
\method{test_connection}{default}(connection)
}
\description{
Default test connection method
}
