% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tide_mon}
\alias{tide_mon}
\alias{tidelevel}
\title{Monthly tide height data from Cochin}
\format{tide_mon is a data frame with:
\describe{
  \item{Year}{The year}
  \item{Month}{The month}
  \item{tide.level.mm}{Tide level in mm above the Cochin datum (1.94m below BM27)}
  \item{tide.level.mm}{Tide level in mm above the Cochin datum (1.94m below BM27)}
  \item{tide.level.interp}{tide.level.mm with the NAs interpolated with exponential smoothing.}
  \item{anomalies}{tide.level.interp with linear trend and monthly means removed.}
  \item{anomalies13}{13 month running mean of anomalies as done in Srinivas et al. 2005 (Fig. 3).}
  \item{anomaliesb}{tide.level.interp with non-linear trend (via gam) and monthly means removed.}
  \item{anomalies13b}{13 month running mean of anomaliesb.}
}}
\usage{
tide_mon
}
\description{
Tide height data were downloaded form three sources to provide a 1939 to 2018
time series of tide height at Cochi. The Cochi tide gauge is on Willington Island; 
\href{https://www.gloss-sealevel.org/gloss-station-handbook?stn=32}{GLOSS# 032}, UH#174, 
\href{http://www.ioc-sealevelmonitoring.org/station.php?code=coch}{IOC#coch}
and \href{https://www.psmsl.org/data/obtaining/stations/438.php}{PSMSL#438}.
The data are the raw tide gauge values, which are the mm above the datum for this gauge.  The datum
is 1.94m below BM27.
  
Sea level measurements are affected by the local barometric pressure, and
it is common to apply the inverted barometer correction which uses
the difference between the current air pressure and the mean air pressure.
However Srinivas and Kumar (2006) found that for the Cochin tide height measurements,
"There is no significant difference between the seasonal march of the observed 
sea level and CSL [corrected sea level]". Therefore we use the raw sea level data with no
correction.
 
Srinivas et al. (2005) and Srinivas and Kumar (2006) found a positive correlation
between the Southern Oscillation Index (SOI) and mean sea level for 1976-1993,
with the sea level decreasing during El Nino years and increasing during La Nina years. 
Singh (2006) also found this correlation but for Sept only for Cochin. A Singh (2006) also found
a correlation between the monthly rainfall in the SW monsoon season and mean sea level. Srinivas
and Kumar (2006) also found a relationship between the along shore winds and sea level height at
Cochi. Note that this relationship between sea level and SOI does not seem to hold for the entire
1939-2018 time series.
}
\details{
For 1939 to 2013, mean monthly tide height was downloaded from the Permanent Service for Mean
Sea Level (PSMSL), (PSMSL 2019, Holgate et al. 2013). This is from the tide gauge at Cochin on 
Willington Island (Lat 9.966667, Lon 76.266667). The data are the raw tide level data, termed
the 'metric' data on the PSMSL website. Download link was 
\url{https://www.psmsl.org/data/obtaining/met.monthly.data/438.metdata}

In 2011, a automatic tide gauge was installed which provides minute tide height data. Data
from 2011 onward are provided by Indian National Centre for Ocean Information Services (INCOIS)
and hosted at \url{http://www.ioc-sealevelmonitoring.org} and the
University of Hawaii Sea Level Center (UHSLC) \url{https://uhslc.soest.hawaii.edu/}. Research quality
data are not available for Cochin (UH# 174). Instead the Fast Delivery data were used
and quality control (outlier removal) was applied to the downloaded data per the algorithm 
used by IOC Sea Level Monitoring (FAQ \url{http://www.ioc-sealevelmonitoring.org/service.php}).

Daily minute tide level data were downloaded from \url{http://www.ioc-sealevelmonitoring.org/station.php?code=coch}
and outlier removal applied.
\itemize{
 \item All values X where abs(X – median) > tolerance are hidden.
 \item With tolerance = 3*abs(percentile90 - median)
 \item The statistics (median and percentile90) are calculated on the daily minute data
 }
 After outlier removal, NAs were replaced with a linear interpolation and 
 the mean tide height for the day (average over all the minute data) was computed. The daily
 data were then averaged over the calendar month to create the monthly average. Months with
 fewer than 20 days of data were given NA.
 was computed.
 
 Daily mean tide level was also downloaded from the
University of Hawaii Sea Level Center (UHSLC) at the link \url{http://uhslc.soest.hawaii.edu/data/?fd#uh174}.
This data should be the same as the IOC data above but there were slight differences that
are likely due to differences in outlier removal and how NAs are treated. As for the IOC data, he daily
 data were then averaged over the calendar month to create the monthly average. Months with
 fewer than 20 days of data were given NA.
 was computed. The reference for these data are Caldwell et al. (2015).
 
 A monthly time series for 1939 to 2018 was assembled by merging the 3 data sets. The PSMSL data
 were used for 1939 to Aug 2011. For Sept 2011 onward, the IOC and UHSLC monthly data were averaged.
 If one had a NA and the other did not, the non-NA value was used. The monthly data had NAs. The NAs
 were interpolated using exponential smoothing (ETS) via a seasonal model. 
 \itemize{
 \item An ETS model was fit to the data up to the first NA
 \item The fitted ETS model was used to forecast one step ahead.
 \item The one-step ahead forecast was used to fill in the first NA
 \item A new ETS model was fit to the data up to the second NA, where the first NA was filled in via the step above.
 \item The new fitted ETS model was used to forecast one step ahead.
 \item This one-step ahead forecast was used to fill in the second NA.
 \item This process was repeated until all NAs were replaced.
 }
 The R code for interpolation in the 'process-merged-tide-gauge-data.R' file.
 
 All code for downloading and processing the data along with the raw downloaded
 data is in the
 extdata/tide gauge data folder in the doc folder.
}
\references{
Caldwell, P. C., M. A. Merrifield, P. R. Thompson (2015), 
Sea level measured by tide gauges from global oceans — the Joint Archive for Sea 
Level holdings (NCEI Accession 0019568), Version 5.5, NOAA National Centers for 
Environmental Information, Dataset, doi:10.7289/V5V40S7W. Research Quality data is not
available for the Cochin tide guage (UH# 174, GLOSS# 032). The Fast Delivery data was downloaded on May 7, 2019.
from \url{http://uhslc.soest.hawaii.edu/data/?fd#uh174}

\insertRef{Holgateetal2013}{SardineForecast}

Permanent Service for Mean Sea Level (PSMSL). 2019. 
"Tide Gauge Data", Retrieved 6 May 2019 from 
\url{https://www.psmsl.org/data/obtaining/stations/438.php}.

\insertRef{Singh2006}{SardineForecast}

\insertRef{SrinivasKumar2006}{SardineForecast}

\insertRef{Srinivasetal2005}{SardineForecast}
}
\keyword{datasets}
