% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tide_yr}
\alias{tide_yr}
\title{Yearly mean tide height data from Cochin
#'}
\format{
tide_yr is a data frame with:
\describe{
  \item{Year}{The year}
  \item{Month}{The month}
  \item{mean}{Calendar year mean of months in tide_mon}
  \item{anomalies}{mean with linear trend and mean level removed.}
}
}
\usage{
tide_yr
}
\description{
See \link{tidelevel} for information on the tide height data and references.
tide_yr is the calendar year averages of the monthly tide levels in tide_mon.
}
\keyword{datasets}
