% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-utils.R
\name{withings_read_csv}
\alias{withings_read_csv}
\title{Read Withings CSV helper}
\usage{
withings_read_csv(path)
}
\arguments{
\item{path}{Path to CSV (epochs or summary)}
}
\value{
tibble
}
\description{
Read Withings CSV helper
}
\keyword{internal}
