% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withings-sleep.R
\name{withings_sleep_report}
\alias{withings_sleep_report}
\title{Generate a concise report for a withings_sleep object}
\usage{
withings_sleep_report(x)
}
\arguments{
\item{x}{withings_sleep}
}
\value{
A tibble summarising key metrics per record
}
\description{
Generate a concise report for a withings_sleep object
}
