% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enso}
\alias{enso}
\title{ENSO Indices}
\format{A data frame with variables:
\describe{
  \item{Year}{}
  \item{Month}{For ONI and SOI, this is a 3-month average.  Jan value is average of Dec-Jan-Feb.  Qtr would be denoted JFM.}
  \item{ONI}{The ONI index value}
  \item{SOI}{The SOI index value}
  \item{DMI}{The DMI index value}
  \item{MEI}{The MEI index value}
}}
\usage{
enso
}
\description{
Oceanic Nino Index, Equatorial Southern Oscillation Index and Dipole Mode Index
}
\details{
The ONI index is 3 month running mean of ERSST.v5 SST anomalies in the Niño 3.4 region 
(5°N-5°S, 120°-170°W)], based on centered 30-year base periods updated every 5 years.  
The ONI was downloaded as follows:
\preformatted{
require(XML)
require(RCurl)
require(httr)
oni=read.table("http://www.cpc.ncep.noaa.gov/data/indices/oni.ascii.txt", stringsAsFactors=FALSE)
}
The SOI used is the Equatorial SOI monthly.  The data were downloaded with the following code.
\preformatted{
soi=read.table("http://www.cpc.ncep.noaa.gov/data/indices/reqsoi.for", stringsAsFactors=FALSE)
}
The DMI  is the monthly Dipole Mode Index.  
The DMI (also IOD index) is defined by the SSTA difference between the 
western Indian Ocean (10°S–10°N, 50°E–70°E) and the southeastern Indian Ocean (10°S–0°, 90°E–110°E).
The data were downloaded with the following code.
\preformatted{
dmi=read.table("https://www.esrl.noaa.gov/psd/gcos_wgsp/Timeseries/Data/dmi.long.data", skip=1, nrows=149)
}
The IOBM index is defined as the SSTA averaged over the tropical Indian Ocean (40°E–100°E, 20°S–20°N).
}
\references{
\insertRef{SajiYamagata2003}{SardineForecast}
}
\keyword{datasets}
