% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loogam.R
\name{loogam}
\alias{loogam}
\title{Leave one out cross-validation on gam or lm object}
\usage{
loogam(mod, k = 1, n = 100)
}
\arguments{
\item{mod}{A model returned by a gam or lm fit}

\item{k}{samples. k=1 is LOO, k=2 is LTO}
}
\value{
A list with predictions (fitted), actual, errors (fitted-actual), MAE and RMSE.
}
\description{
Helper function to do a LOOCV on a gam or lm model
}
\examples{
# some simulated data with one cov
dat <- mgcv::gamSim(6,n=100,scale=.5)[,1:2]
m <- mgcv::gam(y~s(x0), data=dat)
loogam(m)$MAE
}
