% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-connection.R
\name{test_connection}
\alias{test_connection}
\title{Test Connection to Data Source}
\usage{
test_connection(connection, ...)
}
\arguments{
\item{connection}{A sardine_connection object}

\item{...}{Additional arguments passed to source-specific methods}
}
\value{
Logical. TRUE if connection is successful, FALSE otherwise
}
\description{
Generic function to test connections to various data sources. Dispatches
to source-specific testing methods.
}
\examples{
\dontrun{
  conn <- redcap_connection(url = "https://example/api/", token = "TOKEN")
  ok <- test_connection(conn)
}
}
